FlaskApp/requirements.txt
altgraph==0.16.1
appnope==0.1.0
astroid==2.4.1
attrs==19.1.0
autopep8==1.5.2
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.0
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
click==7.1.1
cocos2d==0.6.5
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
et-xmlfile==1.0.1
Flask==1.1.1
Flask-SQLAlchemy==2.5.1
future==0.17.1
greenlet==1.1.0
idna==2.9
importlib-metadata==4.5.0
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.13.3
Jinja2==2.10.1
joblib==0.15.1
json5==0.9.4
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==2.1.2
jupyterlab-server==1.1.3
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
macholib==1.11
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.3
openpyxl==3.0.3
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pefile==2019.4.18
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.0
prompt-toolkit==2.0.9
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pycodestyle==2.6.0
pygame==1.9.6
pyglet==1.3.2
Pygments==2.4.2
PyInstaller==3.4
pylint==2.5.2
pyparsing==2.4.0
PyQt5==5.14.2
PyQt5-sip==12.7.2
PyQt5-stubs==5.14.2.0
pyrsistent==0.15.2
pyserial==3.4
python-dateutil==2.8.0
pytz==2019.1
pyzmq==18.0.1
qtconsole==4.5.1
qtwidgets==0.16
requests==2.23.0
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
soupsieve==2.0
SQLAlchemy==1.4.17
Tcl==0.2
terminado==0.8.2
testpath==0.4.2
threadpoolctl==2.1.0
tk==0.1.0
toml==0.10.1
tornado==6.0.4
traitlets==4.3.2
typed-ast==1.4.1
typing-extensions==3.10.0.0
urllib3==1.25.9
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.4.2
wrapt==1.12.1
zipp==3.4.1
